Ejercicio_1/requirements.txt
# Dependencias del proyecto Ejercicio 1 - ML vs LLM

# Data manipulation
pandas==2.1.4
numpy==1.26.3

# Machine Learning
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# LLM APIs
groq==0.11.0

# Progress bars
tqdm==4.66.1

# Statistical analysis
scipy==1.11.4
